requirements.txt
instalooter==2.4.4
python-telegram-bot==13.7
youtube-dl==2021.6.6